Generation in the world
Soulbound books now spawn in ancient cities, but I should change the weight of them.
SoulboundUtil now has more functions, including giveSoulbound and safeGiveSoulbound, the latter of which is there to fix an error with adding an endless amount of tags when adding soulbound to an item in an anvil.